+++ high_five.rb
@@ -1,0 +1,6 @@
+module HighFive
+  def offer_high_five
+    "High five!"
+  end
+
+end
